Optimize delegation
For _void_element we can directly use `alias_method` and then
use `__callee__.name` to know which tag to generate.

This way it saves a method call, with `...` delegation which
isn't that fast.

Unfortunately `_standard_element` can't be delegated this way
because of the `template` method. However we can enable
`frozen_string_literal: true` in `class_eval` to save a string
allocation per call.