Add Phlex components to Tailwind configuration
Phlex components are Ruby files, so they have a `.rb` extension.

That's why, to be able to use Phlex components with Tailwind, we need
to add the `.rb` extension to the list of files that Tailwind should
look at when generating CSS.

This is done in the `config/tailwind.config.js`:

```js
module.exports = {
  content: [
    // ...
    './app/views/**/*.{erb,haml,html,slim,rb}'
  ]
}
```

Thanks to this commit, we automatically add the `.rb` extension to the
list of extensions when it is not already present.