Use standard elements only for SVG
In HTML, SVG is considered a foreign element and must either have a
start tag and an end tag, or a start tag that is marked as self-closing,
in which case they must not have an end tag. See
https://html.spec.whatwg.org/multipage/syntax.html#foreign-elements for
details. Most (all?) of these SVG elements can have child elements too,
mainly descriptive elements like ‘desc’, ‘metadata’ and ‘title’.