Experimental Smart chunking (#535)
This PR updates Phlex to write to the buffer after `</head>`.
Additionally, it provides a `await` interface for waiting on
asynchronous tasks. If the task is finished, it will continue rendering;
otherwise, it will write to the buffer while waiting.

```ruby
class Articles < Phlex::HTML
  def initialize(articles_query)
    @articles_query = articles_query
  end

  def template
    h1 { "Articles" }

    ul(class: "articles")
      await(@articles_query).each do |article|
        li { article.title }
      end
    end
  end
end
```

In this example, if the `articles_query` isn't finished by the time we
await it, we will write the following to the buffer while waiting for it
to resolve.

```html
<h1>Articles</h1><ul class="articles">
```

The component can be invoked with an Async task like this:

```ruby
Sync do
  Articles.new(
    Async { some_slow_query }
  ).call(stream)
end
```

Or using Concurrent Ruby IVars. For example, if `SomeClass` includes
`Concurrent::Async`, we could invoke the component like this.

```ruby
Articles.new(SomeClass.async.some_slow_query).call
```

In the future, this feature could be patched by phlex-rails to support
asynchronous ActiveRecord queries, e.g.

```ruby
render ArticlesComponent.new(Article.all.load_async)
```