Support for Array and Set attribute values (#546)
This adds support for Array or Set attribute values like this:

```ruby
div(class: ["bg-red-500", "rounded"])
# or
div(class: Set["bg-red-500", "rounded"])
```

Array or Set attribute values will be treated like a token list, and its
members will be joined with a space. This works as long as all elements
are either strings or symbols.

Additionally, this change removes the implicit conversion of attribute
values to a string by calling `to_s` on anything that is not recognized.
Now attribute values that are not either a String, a Symbol, an Array, a
Set, or a Hash with values of one of these types, will raise an
ArgumentError.