Apply consistent symbol/string attribute transformation for Hash values
Symbol attribute keys are dasherized before rendering, whereas string
keys are passed through verbatim.

However if the value is a Hash, the dasherize transform will not be
applied to the key even if it is a symbol. This change makes it so the
dasherization is applied even in this case.

```rb
a(data_turbo: { method: "post", confirm: "Are you sure?" })
```

Should render

```html
<a data-turbo-method="post" data-turbo-confirm="Are you sure?"></a>
```