A different set of trade offs for preventing CSV injections
The goal of this re-prioritization is to lighten the burden on
Phlex::CSV users. I expect that most people will not face a CSV
injection attack, and forcing them to think about it all the time will
likely be met with frustration more than thankfulness that we're making
them be safe.

To that end here are the trades offs presented here:
* All escaped values (column headers, and row values) are checked for
  CSV injection characters. I believe the cost of this (a single hash
  table lookup) is small enough to be worth the improved user experience
* `escape_csv_injection?` is called only when a suspected injection has
  been detected
* `escape_csv_injection?` has its arity checked and cached the first
  time it gets called. This allows users to define a simple override:

    def escape_csv_injection? = true

  But also allows them to have finer grain control if needed

    def escape_csv_injection?(value, column_header)
* Finally our default implementation raises an error message that will
  explain CSV injection attacks, show the suspicious value and which
  column it was in (if we know it), and how to proceed