Render from namespaced component buckets (#676)
This PR makes it possible to render components from external bucket
namespaces. So if `Fancy` is a `Phlex::Bucket` with a `Table` component,
you could render it like this.

```ruby
class UsersTable < Phlex::HTML
  def initialize(users:)
    @users = users
  end

  def view_template
    h2 { "Users" }

    Fancy::Table(@users) do |t|
      t.column "Name", &:name
      t.column "Email", &:email
    end
  end
end
```

### How it works

Phlex uses Fiber storage to keep track of the current rendering context.
When the `Table` component is added to the `Fancy` bucket, a singleton
method is defined on the bucket module with the same name as the
component.

You can also include the bucket to have direct references.

```ruby
class UsersTable < Phlex::HTML
  include Fancy

  def initialize(users:)
    @users = users
  end

  def view_template
    h2 { "Users" }

    Table(@users) do |t|
      t.column "Name", &:name
      t.column "Email", &:email
    end
  end
end
```

### Notable changes

Using Fiber storage is a bit weird. I think it’s fine, but I think we
should make this feature experimental for now, in case we want to revert
it.