Lazy load autoloaded bucket components (#678)
Autoloaded Ruby constants will not have corresponding proxy methods on
Buckets, since these methods are defined on `const_added`. This PR
solves that in two ways:

### External access
Assuming a runtime environment where none of the constants are loaded by
default — such as Rails with Zeitwerk during development — we use a
bucket like this:

```ruby
def view_template
  Components::Heading { "Hello" }
end
```

This call will autoload the `Components` module, but it will not
autoload the `Components::Heading` module, since it’s just a method call
on `Components`. How we handle this is the `Components` bucket catches
the call with `method_missing`, and then tries to load the constant. If
loading the constant is successful, it can call the newly defined
`Heading` method on `Components`.

### When included
If we include `Components` and then try to call `Heading`, again we will
run into the same problem.

```ruby
class MyComponent < Phlex::HTML
  include Components

  def view_template
    Heading { "Hello" }
  end
end
```

In this case, though, we can’t use `method_missing`. If you include
`Components` and `AnotherBucket` and they both define a `Heading`
component, you would expect `AnotherBucket`’s `Heading` component to
take precedence over `Components`’ `Heading` component. If both buckets
were using the `method_missing` technique, this precedence would be
based on which bucket had their component autoloaded first.

So instead, when you _include_ a bucket, we eager-load all that bucket’s
components.