Add `__nested__attributes__`
Special-cases the serialization of nested attributes to avoid
allocations. This avoids having to build up a new hash on each recursive
call, and lets us "bypass" a lot of the safety features (such as
dangerous attribute sanitation) that Phlex includes when serializing
nested attributes on elements.

This should redeuce the number of allocations and make the code more
performant.