Move capturing up to the state object (#880)
In order to fix https://github.com/phlex-ruby/phlex-rails/issues/276, we
need to simplify the capturing interface so it can be swapped out. This
PR removes the ability to capture into an arbitrary buffer. Instead, it
depends on reading the new buffer from the state within a capture block.

See https://github.com/phlex-ruby/phlex-rails/pull/279